requirements.txt
requests==2.31.0
python-chess
pytest==8.1.1